requirements.txt
Flask==2.0.1
opencv-python-headless==4.5.3.56
openvino-dev==2021.4
